[CAZB-3756] Non-alphabetic/numeric characters in Account Search cause Service Unavailable error (#848)
